move ProtocolManager into its own unit
