+++ cvmts/src/protocol/GuacamoleProtocol.ts
@@ -1,5 +1,5 @@
 import pino from 'pino';
-import { IProtocol, IProtocolHandlers, ListEntry, ProtocolAddUser, ProtocolBase, ProtocolChatHistory, ProtocolFlag, ProtocolRenameStatus, ProtocolUpgradeCapability, ScreenRect } from './Protocol.js';
+import { IProtocol, IProtocolMessageHandler, ListEntry, ProtocolAddUser, ProtocolBase, ProtocolChatHistory, ProtocolFlag, ProtocolRenameStatus, ProtocolUpgradeCapability, ScreenRect } from './Protocol.js';
 import { Rank, User } from '../User.js';
 
 import * as cvm from '@cvmts/cvm-rs';
